% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_AUC_expected_QR.R
\name{calc_AUC_expected_QR}
\alias{calc_AUC_expected_QR}
\title{Calculate the expected AUC under the QR model}
\usage{
calc_AUC_expected_QR(
  input_data, timepoint = 12,
  cpeptide_baseline_colname = "auc_mean", age_colname = "age_years")
}
\arguments{
\item{input_data}{data frame containing the C-peptide AUC values from MMTTs and subject ages.}

\item{timepoint}{numeric, the number of months at which to calculate the expected C-peptide value. Current implementation only supports 12.}

\item{cpeptide_baseline_colname}{character or numeric, the column containing the C-peptide values. Defaults to "auc_mean".}

\item{age_colname}{character or numeric, the column containing the age values. Defaults to "age_years".}
}
\value{
a vector containing the expected C-peptide values at the specified timepoint, in ln(nmol/L/min + 1).
}
\description{
This function calculates the expected AUC under the quantitative response
(QR) model for a mixed meal tolerance test (MMTT), following the approach
developed by Bundy et al. 2020 (PMID 32704564), with expansions for other
timepoints by Ylescupidez et al. 2023 (PMID 37940642). As input,
it takes a data frame with age and baseline C-peptide values for each
subject, and a timepoint for which to calculate the expected AUC.
NOTE: Values for AUC must be in the standardized units used in the QR model,
specifically in units of nmol per L per minute from a 120-minute MMTT
(not log-transformed). Values for age must be in years.
}
\details{
The equation used is:
"Cpep_expected = -0.191 + 0.812 * ln(Cpep_baseline + 1) + 0.00638 * Age"
}
