% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Cpeptide_rates.R
\name{calc_Cpeptide_rates}
\alias{calc_Cpeptide_rates}
\title{Estimate rate of C-peptide change over time, by subject (with options!)}
\usage{
\code{
calc_Cpeptide_rates(
  cpeptide_auc_data, model_type=c("independent", "random_effect"),
  identifier_column="subject", time_column="cpeptide_study_day",
  auc_column="auc",
  group_column
  )}
}
\arguments{
\item{cpeptide_auc_data}{data frame containing the C-peptide data. Should contain a unique subject identifer, a numeric column for the timing of visits, and C-peptide AUC values. Any non-NA values included in this data frame will be used, so filtering should be done before passing the data to this function.}

\item{model_type}{character, the type(s) of model to fit. Options are "independent", "random_effect", and "grouped_random_effect". "independent" uses a simple fixed-effects model with slopes and intercepts for each subject. "random_effect" uses a mixed-effects model with subject-level random intercepts and slopes. "grouped_random_effect" uses a mixed-effects model like "random_effect", but with group-level fixed effects for the slopes.}

\item{identifier_column}{character or numeric, the column containing the subject identifiers. Defaults to "subject".}

\item{time_column}{character or numeric, the column with the time variable. Should be numeric and based on a baseline visit, as intercept values only make sense if they are based on a common scale. Default is "cpeptide_study_day".}

\item{auc_column}{character or numeric, the column containing the C-peptide AUC values. Data should be transformed to whatever form the model is to be fit to (typically log-transformed). Defaults to "auc".}

\item{group_column}{character or numeric, the column containing the subject grouping. Used only if \code{model_type} is set to "grouped_random_effect"}
}
\value{
a list containing, for each value of \code{model_type}, an element with the model fit(s) (as $model) and and a data frame with the extracted slopes and intercepts for each subject (as $rates). Slopes are given in the units of \code{auc_column} / unit of \code{time_column}.
}
\description{
This function uses linear models to estimate the rate of C-peptide AUC change over time. 
There are three options for how to fit models, set using the \code{model_type} argument:
"independent" fits a separate, independent intercept and slope for each value of
\code{identifier_column}; "random_effect" fits an individual-level random effect for both the
intercept and the slpope; "grouped_random_effect" fits an individual-level random effect for
the intercept, a group-level fixed effect for each unique value of \code{group_column}, and
an individual-level random effect around the group means.
}
