% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dptrs.R
\name{calc_dptrs}
\alias{calc_dptrs}
\title{Calculate the Diabetes Prevention Trial-Type 1 Risk Score (DPTRS), a composite measure of pre-diagnosis T1D status}
\usage{
calc_dptrs(
  input_data,
  cpeptide_fasting_colname = "pep0",
  cpeptide_sum_colname = NULL,
  cpeptide_colname_prefix = NULL,
  glucose_sum_colname = NULL,
  glucose_colname_prefix = NULL,
  bmi_colname,
  age_colname)
}
\arguments{
\item{input_data}{data frame containing the C-peptide, glucose values, BMI, and age data}

\item{cpeptide_fasting_colname}{character or numeric, the column containing the fasting C-peptide values in ng/mL. Defaults to "cpeptide_fasting".}

\item{cpeptide_sum_colname}{character or numeric, the column containing the sum of OGTT C-peptide values in ng/mL for 30, 60, 90, 120 minutes. Defaults to NULL, and EITHER this or \code{cpeptide_colname_prefix} must be provided, but not both.}

\item{cpeptide_colname_prefix}{character, the name prefix of the columns containing the OGTT C-peptide values in ng/mL for 30, 60, 90, 120 minutes. Defaults to NULL, and EITHER this or \code{cpeptide_sum_colname} must be provided, but not both.}

\item{glucose_sum_colname}{character or numeric, the column containing the sum of OGTT C-peptide values in mg/dL for 30, 60, 90, 120 minutes. Defaults to NULL, and EITHER this or \code{glucose_colname_prefix} must be provided, but not both.}

\item{glucose_colname_prefix}{character, the name prefix of the columns containing the OGTT C-peptide values in mg/dL for 30, 60, 90, 120 minutes. Defaults to NULL, and EITHER this or \code{glucose_sum_colname} must be provided, but not both.}

\item{bmi_colname, }{character or numeric, the column containing the BMI data. Defaults to "bmi"}

\item{age_colname, }{character or numeric, the column containing the age data. Defaults to "age"}
}
\value{
a vector containing the DPTRS values
}
\description{
This function calculates the Diabetes Prevention Trial-Type 1 Risk Score of
Sosenko et al. 2015 (PMID 26077017) and Sosenko et al. 2008 (PMID  18000175),
a composite measure of pre-diagnosis T1D status from an oral glucose
tolerance test (OGTT) that has been found to predict onset of T1D. Published
work indicates that Index60 values above 7.0 or 9.0 may predict onset of
clinical T1D. The function takes an input data frame with fasting C-peptide
levels; glucose data from 30, 60, 90, 120 minutes, C-peptide data from 30,
60, 90, 120 minutes; ; and age.
NOTE: Values for all input data must must be in the standardized units used
in DPTRS, specifically ng/mL for C-peptide, mg/dL for glucose, km/m^2 for
BMI, and years for age.
}
\details{
The equation used is:
"DPTRS = (1.57 * log BMI) − (0.06 * age) + (0.81 * glucose sum from 30 to 120 min/100) − (0.85 * C-peptide sum from 30 to 120 min/10) + (0.48 * log fasting C-peptide)
}
