% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_QR.R
\name{calc_QR}
\alias{calc_QR}
\title{Calculate the quantitative response (QR) of C-peptide AUC}
\usage{
calc_QR(
  input_data, timepoint = 12, cpeptide_baseline_colname = "aucMeanBaseline",
  cpeptide_timepoint_colname = "aucMean", age_colname = "ageYears")
}
\arguments{
\item{input_data}{data frame containing the C-peptide AUC values from MMTTs and subject ages.}

\item{timepoint}{numeric, the number of months at which to calculate the expected C-peptide value. Current implementation only supports 12.}

\item{cpeptide_baseline_colname}{character or numeric, the column containing the C-peptide values. Defaults to "aucMeanBaseline".}

\item{cpeptide_timepoint_colname}{character or numeric, the column containing the C-peptide values. Defaults to "aucMean".}

\item{age_colname}{character or numeric, the column containing the age values. Defaults to "ageYears".}
}
\value{
a vector containing the QR values at the specified timepoint, in ln(nmol/L/min + 1).
}
\description{
This function calculates the quantitative response (QR) of Bundy et al. 2020
(PMID 32704564) for a mixed meal tolerance test (MMTT), with expansions for
other timepoints by Ylescupidez et al. 2023 (PMID 37940642). As input, it
takes a data frame with age, baseline C-peptide AUC, and C-peptide AUC at
the specified timepoint (all for each subject), and the timepoint for which
to calculate the QR.
NOTE: Values for AUC must be in the standardized units used in the QR model,
specifically in units of nmol per L per minute from a 120-minute MMTT
(not log-transformed). Values for age must be in years.
}
\details{
The equation used is:
"QR = Cpep_actual - Cpep_expected = ln(Cpep_timepoint + 1) + 0.191 - 0.812 * ln(Cpep_baseline + 1) - 0.00638 * Age"
}
